Login, register, forgot password with cognito. We have OTP in Email. @Puppy check fix UX\UI Forgot password (http://localhost:4200/forgot-password).

+++ ellora/backend/src/main/java/com/bookingnailms/dto/salon/SalonSummaryResponse.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data
 @Builder
@@ -16,6 +17,6 @@ public class SalonSummaryResponse {
     private String address;
     private String city;
     private String logoUrl;
-    private Double averageRating;
+    private BigDecimal averageRating;
     private Integer totalReviews;
 }
